refactor: 홈 화면 UI 구조 리팩터링 및 컴포넌트 재구성
홈 화면의 복잡한 UI 로직을 분리하고, 재사용성을 높이기 위해 컴포넌트 구조를 전면 개편하였습니다.

* **refactor: 홈 화면 레이아웃 및 히어로 영역 컴포넌트 분리**
    * 기존 `EmptyPresentationContent`, `PresentationContent`를 제거하고, 공통 레이아웃인 `HomePageLayout`으로 통합했습니다.
    * 히어로 영역 UI를 `EmptyPresentationHero`와 `PresentationHero`로 분리하여 각 상태별 책임을 명확히 했습니다.
    * `HomeTitleSection`의 명칭을 `HomtHeroLayout`으로 변경하였습니다.

* **refactor: 바텀 시트 구조 개선 및 관련 컴포넌트 추가**
    * 바텀 시트 내의 콘텐츠를 `EmptyPresentationSheet`와 `PresentationSheet`로 각각 모듈화했습니다.
    * 바텀 시트 전용 타이틀 컴포넌트인 `HomeBottomSheetTitle`을 추가했습니다.
    * `HomeBottomSheetScaffold` 및 `HomeBottomSheetScope` 등 기존의 복잡한 내부 스캐폴드 로직을 정리했습니다.

* **refactor: 공통 유틸리티 이동 및 정리**
    * 높이 변화를 감지하는 `onHeightChanged` 확장 함수를 `feature:home` 모듈에서 공통 `core:ui` 모듈로 이동하여 재사용성을 높였습니다.

* **style: 리소스 참조 및 프리뷰 코드 정리**
    * 변경된 컴포넌트 구조에 맞춰 `HomeScreen`의 렌더링 로직을 업데이트하고, 불필요한 import와 어노테이션을 정리했습니다.
